+++ R/themes.R
@@ -362,51 +362,6 @@ theme_mpr <- function(
       )
 }
 
-
-#' @export
-#' @rdname theme_mpr
-theme_inflation_report <- function(
-  base_family = "Calibri", base_size = 11.5, base_colour = "#1e1e1e",
-  plot_title_face = "plain", plot_title_size = 18, plot_title_colour = "#00294E", 
-  plot_title_position = c("panel", "plot"),
-  subtitle_face = "plain", subtitle_colour = base_colour, subtitle_size = 14,  
-  axis_title_size = base_size, axis_title_colour = base_colour,
-  axis_title_just = "mc",
-  axis_text_size = base_size, axis_text_colour = base_colour,
-  axis_ticks_length = 0.35,
-  legend_title_size = base_size, legend_title_colour = base_colour, 
-  legend_text_size = base_size, legend_text_colour = base_colour, 
-  legend_position = "top", legend_just = "left", 
-  legend_margin = c(base_size/2, base_size/2, base_size/2, 0),
-  caption_face = "plain", caption_colour = base_colour, caption_size = base_size,  
-  strip_text_face = "plain", strip_text_colour = base_colour, 
-  strip_text_size = 13, plot_margin = c(base_size/2, base_size/2, base_size/2, base_size*2)
-  ) {
-    
-  .Deprecated("theme_mpr")
-    
-  theme_mpr(
-    base_family = base_family, base_size = base_size, base_colour = base_colour,
-    plot_title_face = plot_title_face, plot_title_size = plot_title_size, 
-    plot_title_colour = plot_title_colour,
-    plot_title_position = match.arg(plot_title_position),
-    subtitle_face = subtitle_face, subtitle_colour = subtitle_colour, 
-    subtitle_size = subtitle_size,
-    axis_title_size = axis_title_size, axis_title_colour = axis_title_colour,
-    axis_title_just = axis_title_just,
-    axis_text_size = axis_text_size, axis_text_colour = axis_text_colour,
-    axis_ticks_length = axis_ticks_length,
-    legend_title_size = legend_title_size, legend_title_colour = legend_title_colour, 
-    legend_text_size = legend_text_size, legend_text_colour = legend_text_colour, 
-    legend_position = legend_position, legend_just = legend_just, legend_margin = legend_margin,
-    caption_face = caption_face, caption_colour = caption_colour, 
-    caption_size = caption_size,
-    strip_text_face = strip_text_face, strip_text_colour = strip_text_colour, 
-    strip_text_size = strip_text_size, 
-    plot_margin = plot_margin
-    )
-}
-
 #' A \code{ggplot2} theme for Financial Stability Report (FSR) chart styling
 #'
 #' @description Provides a theme to produce